@@ -1,96 +1,95 @@
 {
-    "name": "@akanass/rx-otp",
-    "version": "1.1.0",
-    "description": "HMAC-based (HOTP) and Time-based (TOTP) One-Time Password manager. Works with Google Authenticator for Two-Factor Authentication.",
-    "main": "commonjs/index.js",
-    "module": "esm5/index.js",
-    "es2015": "esm2015/index.js",
-    "types": "index.d.ts",
-    "private": false,
-    "scripts": {
-        "pretest": "make clean && make pretest",
-        "test": "make test",
-        "coveralls": "make coveralls",
-        "clean": "make clean",
-        "packaging": "make packaging",
-        "browserify": "make browserify",
-        "prebuild": "make clean && make pretest && make test",
-        "build": "make commonjs && make esm2015 && make esm5",
-        "postbuild": "make packaging && make browserify"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git@github.com:akanass/rx-otp.git"
-    },
-    "keywords": [
-        "One-Time Password",
-        "OTP",
-        "HOTP",
-        "RFC 4226",
-        "TOTP",
-        "RFC 6238",
-        "Google Authenticator",
-        "Two-Factor Authentication",
-        "RFC 3548",
-        "RxJS",
-        "Rx",
-        "ReactiveX",
-        "Observable",
-        "Observer",
-        "Streams",
-        "Async",
-        "ES2015",
-        "ES2016",
-        "ES2017",
-        "ES2018",
-        "ES6",
-        "ES7",
-        "ES8",
-        "ES9",
-        "Typescript",
-        "Browser"
-    ],
-    "author": {
-        "name": "Nicolas Jessel",
-        "email": "akanass@icloud.com",
-        "url": "https://github.com/akanass"
-    },
-    "license": "SEE LICENSE IN https://github.com/akanass/rx-otp/blob/master/LICENSE.md",
-    "bugs": {
-        "url": "https://github.com/akanass/rx-otp/issues"
-    },
-    "homepage": "https://github.com/akanass/rx-otp#readme",
-    "dependencies": {
-        "@types/node": "^12.6.2",
-        "@types/qr-image": "^3.2.1",
-        "@types/uid-safe": "^2.1.2",
-        "ajv": "^6.10.1",
-        "ajv-keywords": "^3.4.1",
-        "big-integer": "^1.6.44",
-        "convert-base": "^0.1.0",
-        "pad-component": "^0.0.1",
-        "qr-image": "^3.2.0",
-        "thirty-two": "^1.0.2"
-    },
-    "devDependencies": {
-        "@types/fs-extra": "^8.0.0",
-        "@types/jest": "^24.0.15",
-        "browserify": "^16.3.0",
-        "coveralls": "^3.0.5",
-        "exorcist": "^1.0.1",
-        "fs-extra": "^8.1.0",
-        "jest": "^24.8.0",
-        "rimraf": "^2.6.3",
-        "rxjs": "^6.5.2",
-        "ts-jest": "^24.0.2",
-        "ts-node": "^8.3.0",
-        "tslint": "^5.18.0",
-        "typescript": "^3.5.3"
-    },
-    "peerDependencies": {
-        "rxjs": "^6.5.2"
-    },
-    "engines": {
-        "node": ">=7.0.0"
-    }
+  "name": "@akanass/rx-otp",
+  "version": "2.0.0",
+  "description": "HMAC-based (HOTP) and Time-based (TOTP) One-Time Password manager. Works with Google Authenticator for Two-Factor Authentication.",
+  "main": "commonjs/index.js",
+  "module": "esm2015/index.js",
+  "types": "index.d.ts",
+  "private": false,
+  "scripts": {
+    "pretest": "make clean && make pretest",
+    "test": "make test",
+    "coveralls": "make coveralls",
+    "clean": "make clean",
+    "packaging": "make packaging",
+    "prebuild": "make clean && make pretest && make test",
+    "build": "make commonjs && make esm2015",
+    "postbuild": "make packaging"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:akanass/rx-otp.git"
+  },
+  "keywords": [
+    "One-Time Password",
+    "OTP",
+    "HOTP",
+    "RFC 4226",
+    "TOTP",
+    "RFC 6238",
+    "Google Authenticator",
+    "Two-Factor Authentication",
+    "RFC 3548",
+    "RxJS",
+    "Rx",
+    "ReactiveX",
+    "Observable",
+    "Observer",
+    "Streams",
+    "Async",
+    "ES2015",
+    "ES2016",
+    "ES2017",
+    "ES2018",
+    "ES6",
+    "ES7",
+    "ES8",
+    "ES9",
+    "Typescript",
+    "Browser"
+  ],
+  "author": {
+    "name": "Nicolas Jessel",
+    "email": "akanass@icloud.com",
+    "url": "https://github.com/akanass"
+  },
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/akanass/rx-otp/issues"
+  },
+  "homepage": "https://github.com/akanass/rx-otp#readme",
+  "dependencies": {
+    "@types/node": "^16.10.3",
+    "@types/qr-image": "^3.2.4",
+    "@types/uid-safe": "^2.1.2",
+    "ajv": "^8.6.3",
+    "ajv-keywords": "^5.0.0",
+    "big-integer": "^1.6.49",
+    "convert-base": "^0.1.0",
+    "pad-component": "^0.0.1",
+    "qr-image": "^3.2.0",
+    "thirty-two": "^1.0.2"
+  },
+  "devDependencies": {
+    "@types/fs-extra": "^9.0.13",
+    "@types/jest": "^27.0.2",
+    "coveralls": "^3.1.1",
+    "fs-extra": "^10.0.0",
+    "jest": "^27.2.5",
+    "rimraf": "^3.0.2",
+    "rxjs": "^7.4.0",
+    "ts-jest": "^27.0.5",
+    "ts-node": "^10.2.1",
+    "tslint": "^6.1.3",
+    "typescript": "^4.4.3"
+  },
+  "peerDependencies": {
+    "rxjs": "^7.4.0"
+  },
+  "engines": {
+    "node": ">=7.0.0"
+  }
 }
